+++ lib/src/views/camera/camera_preview_components/save_to_gallery.dart
@@ -1,19 +1,23 @@
 import 'dart:async';
 import 'dart:typed_data';
+import 'package:clock/clock.dart';
+import 'package:drift/drift.dart' show Value;
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:twonly/globals.dart';
+import 'package:twonly/src/database/twonly.db.dart';
 import 'package:twonly/src/services/mediafiles/mediafile.service.dart';
 import 'package:twonly/src/utils/misc.dart';
 
 class SaveToGalleryButton extends StatefulWidget {
   const SaveToGalleryButton({
-    required this.storeImageAsOriginal,
     required this.isLoading,
     required this.displayButtonLabel,
     required this.mediaService,
+    this.storeImageAsOriginal,
     super.key,
   });
-  final Future<Uint8List?> Function() storeImageAsOriginal;
+  final Future<Uint8List?> Function()? storeImageAsOriginal;
   final bool displayButtonLabel;
   final MediaFileService mediaService;
   final bool isLoading;
@@ -44,8 +48,32 @@ class SaveToGalleryButtonState extends State<SaveToGalleryButton> {
                 _imageSaving = true;
               });
 
-              await widget.storeImageAsOriginal();
-              await widget.mediaService.storeMediaFile();
+              if (widget.storeImageAsOriginal != null) {
+                await widget.storeImageAsOriginal!();
+              }
+
+              final newMediaFile = await twonlyDB.mediaFilesDao.insertMedia(
+                MediaFilesCompanion(
+                  type: Value(widget.mediaService.mediaFile.type),
+                  createdAt: Value(clock.now()),
+                  stored: const Value(true),
+                ),
+              );
+
+              if (newMediaFile != null) {
+                final newService = MediaFileService(newMediaFile);
+
+                if (widget.mediaService.tempPath.existsSync()) {
+                  widget.mediaService.tempPath.copySync(
+                    newService.tempPath.path,
+                  );
+                } else if (widget.mediaService.originalPath.existsSync()) {
+                  widget.mediaService.originalPath.copySync(
+                    newService.originalPath.path,
+                  );
+                }
+                await newService.storeMediaFile();
+              }
 
               setState(() {
                 _imageSaved = true;